+++ LoftyCAD.h
@@ -144,10 +144,10 @@ extern SCALED scaled_dirn;
 extern SCALED scaled;
 extern float total_angle;
 extern float effective_angle;
-extern float eff_sx;
-extern float eff_sy;
-extern float eff_sz;
-extern float halo_rad;
+extern double eff_sx;
+extern double eff_sy;
+extern double eff_sz;
+extern double halo_rad;
 extern BOOL suppress_drawing;
 extern float bed_xmin;
 extern float bed_ymin;
@@ -360,9 +360,9 @@ int edge_direction(Edge * e, Plane * pl);
 void point_direction(Point * p0, Point * p1, Plane * pl);
 void project(Plane * ap, Plane * princ, Plane * proj);
 float path_total_length(Object * obj);
-BOOL path_tangent_to_intersect(Object * obj, Plane * pl, Bbox *ebox, Plane * tangent, float *ret_len);
+BOOL path_tangent_to_intersect(Object * obj, Plane * pl, Bbox *ebox, Plane * tangent, double*ret_len);
 BOOL path_is_closed(Object* obj);
-int path_subdivide(Object* obj, Plane* initial_tangent, Bbox *ebox, float initial_len, Plane** tangents);
+int path_subdivide(Object* obj, Plane* initial_tangent, Bbox *ebox, double initial_len, Plane** tangents);
 
 
 // Neighbourhood search and picking (neighbourhood.c)
