+++ triangulate.h
@@ -10,13 +10,13 @@
 // Bounding boxes
 void clear_bbox(Bbox *box);
 void expand_bbox(Bbox *box, Point *p);
-void expand_bbox_coords(Bbox* box, float x, float y, float z);
+void expand_bbox_coords(Bbox* box, double x, double y, double z);
 void union_bbox(Bbox *box1, Bbox *box2, Bbox *u);
 BOOL intersects_bbox(Bbox *box1, Bbox *box2);
 BOOL in_bbox(Point* pt, Bbox* box, float tol);
 
 // Regenerate a view list
-void invalidate_all_view_lists(Object *parent, Object *obj, float dx, float dy, float dz);
+void invalidate_all_view_lists(Object *parent, Object *obj, double dx, double dy, double dz);
 void gen_view_list_face(Face *face);
 void update_view_list_2D(Face *face);
 void gen_view_list_arc(ArcEdge *ae);
